changes commands.sh to not generate files if not needed with reduced output

+++ commands.sh
@@ -2,6 +2,9 @@
 mkdir -p /tmp/results/
 export HOST=`hostname`
 export OUTFILE=`mktemp /tmp/results/${HOST}_XXXXXXXX.json`
+rm ${OUTFILE}	# Let the bsp_prototype actually crate hte file if it wants it.
 /home/docker/bsp_prototype "$@" ${OUTFILE}
-mv ${OUTFILE} /results/
+if [ -f ${OUTFILE} ]; then
+	mv ${OUTFILE} /results/
+fi
 exit "$?"
